Refactored! changed event list by current time
